Backport from 2.0 - DataVersion update
